requirements.txt
# 核心依赖
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
pyyaml>=6.0

# 数据处理
pandas>=2.0.0
sqlalchemy>=2.0.0
sqlite-utils>=3.30

# Web框架（后续使用）
flask>=2.3.0
flask-restful>=0.3.9

# 测试
pytest>=7.3.0
pytest-cov>=4.1.0

# 工具
schedule>=1.2.0
tqdm>=4.65.0